send message whatsapp respond to request

+++ database/migrations/2022_03_03_112844_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
 
 //            $table->string('roles');
             $table->integer('role_id');
-            $table->integer('mobile');
+            $table->bigInteger('mobile');
             $table->integer('office_id')->nullable()->unsigned();
             $table->integer('age');
             $table->boolean('is_active')->default(true);
